migrate remaining controller tests to pytest
