+++ ckanext/unhcr/tests/nose/test_controllers.py
@@ -979,794 +979,3 @@ class TestDepositedDatasetController(base.FunctionalTestBase):
         assert_in('submitted dataset', resp.body)
         assert_in('assigned', resp.body)
         assert_in('as Curator', resp.body)
-
-
-class TestUserController(base.FunctionalTestBase):
-
-    def test_sysadmin_not_authorized(self):
-        user = core_factories.User()
-        env = {'REMOTE_USER': user['name'].encode('ascii')}
-        self.app.post('/user/sysadmin', {}, extra_environ=env, status=403)
-
-    def test_sysadmin_invalid_user(self):
-        user = core_factories.Sysadmin()
-        env = {'REMOTE_USER': user['name'].encode('ascii')}
-        self.app.post(
-            '/user/sysadmin',
-            {'id': 'fred', 'status': '1' },
-            extra_environ=env,
-            status=404
-        )
-
-    def test_sysadmin_promote_success(self):
-        admin = core_factories.Sysadmin()
-        env = {'REMOTE_USER': admin['name'].encode('ascii')}
-
-        # create a normal user
-        user = core_factories.User(fullname='Alice')
-
-        # promote them
-        resp = self.app.post(
-            '/user/sysadmin',
-            {'id': user['id'], 'status': '1' },
-            extra_environ=env,
-            status=302
-        )
-        resp2 = resp.follow(extra_environ=env, status=200)
-        assert_in(
-            'Promoted Alice to sysadmin',
-            resp2.body
-        )
-
-        # now they are a sysadmin
-        userobj = model.User.get(user['id'])
-        assert_equals(True, userobj.sysadmin)
-
-    def test_sysadmin_revoke_success(self):
-        admin = core_factories.Sysadmin()
-        env = {'REMOTE_USER': admin['name'].encode('ascii')}
-
-        # create another sysadmin
-        user = core_factories.Sysadmin(fullname='Bob')
-
-        # revoke their status
-        resp = self.app.post(
-            '/user/sysadmin',
-            {'id': user['id'], 'status': '0' },
-            extra_environ=env,
-            status=302
-        )
-        resp2 = resp.follow(extra_environ=env, status=200)
-        assert_in(
-            'Revoked sysadmin permission from Bob',
-            resp2.body
-        )
-
-        # now they are not a sysadmin any more
-        userobj = model.User.get(user['id'])
-        assert_equals(False, userobj.sysadmin)
-
-
-class TestAdminController(base.FunctionalTestBase):
-    def test_index_sysadmin(self):
-        user = core_factories.Sysadmin()
-        env = {'REMOTE_USER': user['name'].encode('ascii')}
-        self.app.get('/ckan-admin', extra_environ=env, status=200)
-
-    def test_index_not_authorized(self):
-        user = core_factories.User()
-        env = {'REMOTE_USER': user['name'].encode('ascii')}
-        self.app.get('/ckan-admin', extra_environ=env, status=403)
-
-
-class TestExtendedPackageController(base.FunctionalTestBase):
-
-    # Config
-
-    def setup(self):
-        super(TestExtendedPackageController, self).setup()
-
-        # Users
-        self.sysadmin = core_factories.Sysadmin(name='sysadmin', id='sysadmin')
-        self.user1 = core_factories.User(name='user1', id='user1')
-        self.user2 = core_factories.User(name='user2', id='user2')
-        self.user3 = core_factories.User(name='user3', id='user3')
-
-        # Containers
-        self.container1 = factories.DataContainer(
-            name='container1',
-            id='container1',
-            users=[
-                {'name': 'user1', 'capacity': 'admin'},
-            ],
-        )
-        self.container2 = factories.DataContainer(
-            name='container2',
-            id='container2',
-            users=[
-                {'name': 'user2', 'capacity': 'admin'},
-            ],
-        )
-
-        # Datasets
-        self.dataset1 = factories.Dataset(
-            name='dataset1',
-            title='Test Dataset 1',
-            owner_org='container1',
-            data_collection_technique = 'f2f',
-            sampling_procedure = 'nonprobability',
-            operational_purpose_of_data = 'cartography',
-            user=self.user1,
-            visibility='private',
-        )
-
-        # Resources
-        self.resource1 = factories.Resource(
-            name='resource1',
-            package_id='dataset1',
-            url_type='upload',
-            upload=mocks.FakeFileStorage(),
-            url = "http://fakeurl/test.txt",
-        )
-
-    # Helpers
-
-    def make_dataset_request(self, dataset_id=None, user=None, **kwargs):
-        url = '/dataset/copy/%s' % dataset_id
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = self.app.get(url=url, extra_environ=env, **kwargs)
-        return resp
-
-    def make_resource_copy_request(self, dataset_id=None, resource_id=None, user=None, **kwargs):
-        url = '/dataset/%s/resource_copy/%s' % (dataset_id, resource_id)
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = self.app.get(url=url, extra_environ=env, **kwargs)
-        return resp
-
-    def make_resource_download_request(self, dataset_id, resource_id, user=None, **kwargs):
-        url = toolkit.url_for(
-            controller='package',
-            action='resource_download',
-            id=dataset_id,
-            resource_id=resource_id
-        )
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = self.app.get(url=url, extra_environ=env, **kwargs)
-        return resp
-
-    def make_request_access_request(self, dataset_id, user, message, **kwargs):
-        url = '/dataset/{}/request_access'.format(dataset_id)
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = self.app.post(
-            url,
-            {'message': message},
-            extra_environ=env,
-            **kwargs
-        )
-        return resp
-
-    # Dataset
-
-    def test_dataset_copy(self):
-        resp = self.make_dataset_request(dataset_id='dataset1', user='user1')
-        assert_in('action="/dataset/new"', resp.body)
-        #  assert_in('dataset1-copy', resp.body)
-        assert_in('f2f', resp.body)
-        assert_in('nonprobability', resp.body)
-        assert_in('cartography', resp.body)
-        assert_in('Add Data', resp.body)
-        assert_in('container1', resp.body)
-
-    def test_dataset_copy_to_other_org(self):
-        resp = self.make_dataset_request(dataset_id='dataset1', user='user2')
-        assert_in('action="/dataset/new"', resp.body)
-        #  assert_in('dataset1-copy', resp.body)
-        assert_in('f2f', resp.body)
-        assert_in('nonprobability', resp.body)
-        assert_in('cartography', resp.body)
-        assert_in('Add Data', resp.body)
-        assert_not_in('container1', resp.body)
-
-    def test_dataset_copy_no_orgs(self):
-        resp = self.make_dataset_request(dataset_id='dataset1', user='user3', status=403)
-
-    def test_dataset_copy_bad_dataset(self):
-        resp = self.make_dataset_request(dataset_id='bad', user='user1', status=404)
-
-    # Resource Upload
-
-    def test_edit_resource_works(self):
-        url = toolkit.url_for(
-            controller='package',
-            action='resource_edit',
-            id=self.dataset1['id'],
-            resource_id=self.resource1['id']
-        )
-        env = {'REMOTE_USER': self.sysadmin['name'].encode('ascii')}
-
-        # Mock a resource edit payload
-        data = {
-            'id': self.resource1['id'],
-            'name': self.resource1['name'],
-            'type': self.resource1['type'],
-            'description': 'updated',
-            'format': self.resource1['format'],
-            'file_type': self.resource1['file_type'],
-            'date_range_start': self.resource1['date_range_start'],
-            'date_range_end': self.resource1['date_range_end'],
-            'version': self.resource1['version'],
-            'process_status': self.resource1['process_status'],
-            'identifiability': self.resource1['identifiability'],
-
-            'url': 'test.txt',
-            'save': ''
-
-        }
-
-        resp = self.app.post(url, data, extra_environ=env)
-
-        assert 'The form contains invalid entries:' not in resp.body
-
-    def test_edit_resource_must_provide_upload(self):
-        url = toolkit.url_for(
-            controller='package',
-            action='resource_edit',
-            id=self.dataset1['id'],
-            resource_id=self.resource1['id']
-        )
-        env = {'REMOTE_USER': self.sysadmin['name'].encode('ascii')}
-
-        # Mock a resource edit payload
-        data = {
-            'id': self.resource1['id'],
-            'name': self.resource1['name'],
-            'type': self.resource1['type'],
-            'description': 'updated',
-            'format': self.resource1['format'],
-            'file_type': self.resource1['file_type'],
-            'date_range_start': self.resource1['date_range_start'],
-            'date_range_end': self.resource1['date_range_end'],
-            'version': self.resource1['version'],
-            'process_status': self.resource1['process_status'],
-            'identifiability': self.resource1['identifiability'],
-
-            'url': '',
-            'clear_upload': 'true',
-            'save': ''
-
-        }
-
-        resp = self.app.post(url, data, extra_environ=env)
-
-        assert 'The form contains invalid entries:' in resp.body
-        assert 'All data resources require an uploaded file' in resp.body
-
-    # Resource Copy
-
-    def test_resource_copy(self):
-        resp = self.make_resource_copy_request(
-            dataset_id='dataset1', resource_id=self.resource1['id'], user='user1')
-        assert_in('action="/dataset/new_resource/dataset1"', resp.body)
-        assert_in('resource1 (copy)', resp.body)
-        assert_in('anonymized_public', resp.body)
-        assert_in('Add', resp.body)
-
-    def test_resource_copy_no_access(self):
-        resp = self.make_resource_copy_request(
-            dataset_id='dataset1', resource_id=self.resource1['id'], user='user2', status=403)
-
-    def test_resource_copy_bad_resource(self):
-        resp = self.make_resource_copy_request(
-            dataset_id='dataset1', resource_id='bad', user='user1', status=404)
-
-    # Resource Download
-
-    def test_resource_download_anonymous(self):
-        resp = self.make_resource_download_request(
-            dataset_id='dataset1', resource_id=self.resource1['id'], user=None,
-            status=403
-        )
-
-    def test_resource_download_no_access(self):
-        resp = self.make_resource_download_request(
-            dataset_id='dataset1', resource_id=self.resource1['id'], user='user3',
-            status=403
-        )
-
-    def test_resource_download_collaborator(self):
-        core_helpers.call_action(
-            'dataset_collaborator_create',
-            id='dataset1',
-            user_id='user3',
-            capacity='member',
-        )
-        resp = self.make_resource_download_request(
-            dataset_id='dataset1', resource_id=self.resource1['id'], user='user3',
-            status=200
-        )
-
-    def test_resource_download_bad_resource(self):
-        resp = self.make_resource_download_request(
-            dataset_id='dataset1', resource_id='bad', user='user1',
-            status=404
-        )
-
-    def test_resource_download_valid(self):
-        sql = select([
-            model.Activity
-        ]).where(
-            and_(
-                model.Activity.activity_type == 'download resource',
-                model.Activity.object_id == self.dataset1['id'],
-                model.Activity.user_id == 'user1',
-            )
-        )
-
-        # before we start, this user has never downloaded this resource before
-        result = model.Session.execute(sql).fetchall()
-        assert_equals(0, len(result))
-
-        resp = self.make_resource_download_request(
-            dataset_id='dataset1', resource_id=self.resource1['id'], user='user1',
-            status=200
-        )
-
-        # after we've downloaded the resource, we should also
-        # have also logged a 'download resource' action for this user/resource
-        result = model.Session.execute(sql).fetchall()
-        assert_equals(1, len(result))
-
-    # Request Access
-
-    def test_request_access_invalid_method(self):
-        resp = self.app.get(
-            '/dataset/dataset1/request_access',
-            extra_environ={'REMOTE_USER': 'user3'},
-            status=404
-        )
-
-    def test_request_access_missing_message(self):
-        self.make_request_access_request(
-            dataset_id='dataset1', user='user3', message='',
-            status=400
-        )
-
-    def test_request_access_duplicate(self):
-        rec = AccessRequest(
-            user_id=self.user3['id'],
-            object_id=self.dataset1['id'],
-            object_type='package',
-            message='I can haz access?',
-            role='member',
-        )
-        model.Session.add(rec)
-        model.Session.commit()
-        resp = self.make_request_access_request(
-            dataset_id='dataset1', user='user3', message='me again',
-            status=400
-        )
-
-    def test_request_access_invalid_dataset(self):
-        self.make_request_access_request(
-            dataset_id='bad', user='user3', message='I can haz access?',
-            status=404
-        )
-
-    def test_request_access_not_authorized(self):
-        self.make_request_access_request(
-            dataset_id='dataset1', user=None, message='I can haz access?',
-            status=403
-        )
-
-    def test_request_access_valid(self):
-        mock_mailer = mock.Mock()
-        with mock.patch('ckan.plugins.toolkit.enqueue_job', mock_mailer):
-            resp = self.make_request_access_request(
-                dataset_id='dataset1', user='user3', message='I can haz access?',
-                status=302
-            )
-
-        mock_mailer.assert_called_once()
-        assert_equals('user1', mock_mailer.call_args[0][1][0])
-        assert_equals(
-            '[UNHCR RIDL] - Request for access to dataset: "dataset1"',
-            mock_mailer.call_args[0][1][1]
-        )
-        # call_args[0][1][2] is the HTML message body
-        # but we're not going to make any assertions about it here
-        # see the mailer tests for this
-
-        assert_equals(
-            1,
-            len(model.Session.query(AccessRequest).filter(
-                AccessRequest.object_id == self.dataset1['id'],
-                AccessRequest.user_id == self.user3['id'],
-                AccessRequest.status == 'requested'
-            ).all())
-        )
-
-        resp2 = resp.follow(extra_environ={'REMOTE_USER': 'user3'}, status=200)
-        assert_in(
-            'Requested access to download resources from Test Dataset 1',
-            resp2.body
-        )
-
-    def test_request_access_user_already_has_access(self):
-        mock_mailer = mock.Mock()
-        with mock.patch('ckan.plugins.toolkit.enqueue_job', mock_mailer):
-            resp = self.make_request_access_request(
-                dataset_id='dataset1', user='user1', message='I can haz access?',
-                status=302
-            )
-
-        mock_mailer.assert_not_called()
-
-        assert_equals(
-            0,
-            len(model.Session.query(AccessRequest).filter(
-                AccessRequest.object_id == self.dataset1['id'],
-                AccessRequest.user_id == self.user1['id'],
-                AccessRequest.status == 'requested'
-            ).all())
-        )
-
-        resp2 = resp.follow(extra_environ={'REMOTE_USER': 'user1'}, status=200)
-        assert_in(
-            'You already have access to download resources from Test Dataset 1',
-            resp2.body
-        )
-
-
-class TestDataContainer(base.FunctionalTestBase):
-
-    def setup(self):
-        super(TestDataContainer, self).setup()
-
-        self.deposit = factories.DataContainer(
-            name='data-deposit',
-            id='data-deposit',
-        )
-        self.user = core_factories.User(name='user1')
-        self.admin = core_factories.User(name='admin')
-        self.container = factories.DataContainer(
-            name='container1',
-            title='Test Container',
-            users=[
-                {'name': self.admin['name'], 'capacity': 'admin'},
-            ]
-        )
-
-    def make_request_access_request(self, container_id, user, message, **kwargs):
-        url = '/data-container/{}/request_access'.format(container_id)
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = self.app.post(
-            url,
-            {'message': message},
-            extra_environ=env,
-            **kwargs
-        )
-        return resp
-
-    # Request Access
-
-    def test_request_access_invalid_method(self):
-        resp = self.app.get(
-            '/data-container/container1/request_access',
-            extra_environ={'REMOTE_USER': 'user1'},
-            status=404
-        )
-
-    def test_request_access_missing_message(self):
-        self.make_request_access_request(
-            container_id='container1', user='user1', message='',
-            status=400
-        )
-
-    def test_request_access_duplicate(self):
-        rec = AccessRequest(
-            user_id=self.user['id'],
-            object_id=self.container['id'],
-            object_type='organization',
-            message='I can haz access?',
-            role='member',
-        )
-        model.Session.add(rec)
-        model.Session.commit()
-        resp = self.make_request_access_request(
-            container_id='container1', user='user1', message='me again',
-            status=400
-        )
-
-    def test_request_access_invalid_containers(self):
-        # this container doesn't exist
-        self.make_request_access_request(
-            container_id='bad', user='user1', message='I can haz access?',
-            status=404
-        )
-
-        # we can't request access to the data-deposit
-        # because it is _special and different_
-        self.make_request_access_request(
-            container_id='data-deposit', user='user3', message='I can haz access?',
-            status=403
-        )
-
-    def test_request_access_not_authorized(self):
-        self.make_request_access_request(
-            container_id='container1', user=None, message='I can haz access?',
-            status=403
-        )
-
-    def test_request_access_valid(self):
-        mock_mailer = mock.Mock()
-        with mock.patch('ckan.plugins.toolkit.enqueue_job', mock_mailer):
-            resp = self.make_request_access_request(
-                container_id='container1', user='user1', message='I can haz access?',
-                status=302
-            )
-
-        mock_mailer.assert_called_once()
-        assert_equals('admin', mock_mailer.call_args[0][1][0])
-        assert_equals(
-            '[UNHCR RIDL] - Request for access to container: "Test Container"',
-            mock_mailer.call_args[0][1][1]
-        )
-        # call_args[0][1][2] is the HTML message body
-        # but we're not going to make any assertions about it here
-        # see the mailer tests for this
-
-        assert_equals(
-            1,
-            len(model.Session.query(AccessRequest).filter(
-                AccessRequest.object_id == self.container['id'],
-                AccessRequest.user_id == self.user['id'],
-                AccessRequest.status == 'requested'
-            ).all())
-        )
-
-        resp2 = resp.follow(extra_environ={'REMOTE_USER': 'user1'}, status=200)
-        assert_in(
-            'Requested access to container Test Container',
-            resp2.body
-        )
-
-    def test_request_access_user_already_has_access(self):
-        mock_mailer = mock.Mock()
-        with mock.patch('ckan.plugins.toolkit.enqueue_job', mock_mailer):
-            resp = self.make_request_access_request(
-                container_id='container1', user='admin', message='I can haz access?',
-                status=302
-            )
-
-        mock_mailer.assert_not_called()
-
-        assert_equals(
-            0,
-            len(model.Session.query(AccessRequest).filter(
-                AccessRequest.object_id == self.container['id'],
-                AccessRequest.user_id == self.admin['id'],
-                AccessRequest.status == 'requested'
-            ).all())
-        )
-
-        resp2 = resp.follow(extra_environ={'REMOTE_USER': 'admin'}, status=200)
-        assert_in(
-            'You are already a member of Test Container',
-            resp2.body
-        )
-
-
-class TestAccessRequests(base.FunctionalTestBase):
-    def setup(self):
-        super(TestAccessRequests, self).setup()
-
-        self.sysadmin = core_factories.Sysadmin()
-        self.requesting_user = core_factories.User()
-        self.standard_user = core_factories.User()
-        self.pending_user = factories.ExternalUser(state=model.State.PENDING)
-
-        self.container1_admin = core_factories.User()
-        self.container1 = factories.DataContainer(
-            users=[{"name": self.container1_admin["name"], "capacity": "admin"}]
-        )
-        self.container2 = factories.DataContainer()
-        self.dataset1 = factories.Dataset(
-            owner_org=self.container1["id"], visibility="private"
-        )
-        self.container1_request = AccessRequest(
-            user_id=self.requesting_user["id"],
-            object_id=self.container1["id"],
-            object_type="organization",
-            message="",
-            role="member",
-        )
-        self.container2_request = AccessRequest(
-            user_id=self.requesting_user["id"],
-            object_id=self.container2["id"],
-            object_type="organization",
-            message="",
-            role="member",
-        )
-        self.dataset_request = AccessRequest(
-            user_id=self.requesting_user["id"],
-            object_id=self.dataset1["id"],
-            object_type="package",
-            message="",
-            role="member",
-        )
-        self.user_request_container1 = AccessRequest(
-            user_id=self.pending_user["id"],
-            object_id=self.pending_user["id"],
-            object_type="user",
-            message="",
-            role="member",
-            data={'default_containers': [self.container1["id"]]},
-        )
-        self.user_request_container2 = AccessRequest(
-            user_id=self.pending_user["id"],
-            object_id=self.pending_user["id"],
-            object_type="user",
-            message="",
-            role="member",
-            data={'default_containers': [self.container2["id"]]},
-        )
-        model.Session.add(self.container1_request)
-        model.Session.add(self.container2_request)
-        model.Session.add(self.dataset_request)
-        model.Session.add(self.user_request_container1)
-        model.Session.add(self.user_request_container2)
-        model.Session.commit()
-
-    def make_action_request(self, action, request_id, user=None, data=None, **kwargs):
-        url = '/access-requests/{action}/{request_id}'.format(
-            action=action, request_id=request_id
-        )
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = self.app.post(url, data, extra_environ=env, **kwargs)
-        return resp
-
-    def make_list_request(self, user=None, **kwargs):
-        url = '/dashboard/requests'
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = self.app.get(url=url, extra_environ=env, **kwargs)
-        return resp
-
-    def test_access_requests_reject_missing_param(self):
-        self.make_action_request(
-            action='reject',
-            request_id=self.container1_request.id,
-            user=self.container1_admin["name"],
-            status=400,
-            data={},
-        )
-
-    def test_access_requests_invalid_id(self):
-        for action, data in [("approve", {}), ("reject", {'message': 'nope'})]:
-            self.make_action_request(
-                action=action,
-                request_id='invalid-id',
-                user=self.container1_admin["name"],
-                status=404,
-                data=data,
-            )
-
-    def test_access_requests_invalid_user(self):
-        for action, data in [("approve", {}), ("reject", {'message': 'nope'})]:
-            for user in [None, self.standard_user["name"]]:
-                self.make_action_request(
-                    action=action,
-                    request_id=self.container1_request.id,
-                    user=user,
-                    status=403,
-                    data=data,
-                )
-
-    def test_access_requests_approve_container_admin(self):
-        mock_mailer = mock.Mock()
-        with mock.patch('ckanext.unhcr.mailer.mail_user_by_id', mock_mailer):
-            resp = self.make_action_request(
-                action='approve',
-                request_id=self.container1_request.id,
-                user=self.container1_admin["name"],
-                status=302,
-                data={},
-            )
-        mock_mailer.assert_called_once()
-        # standard 'you've been added to a container' email
-        assert_equals(
-            '[UNHCR RIDL] Membership: {}'.format(self.container1['title']),
-            mock_mailer.call_args[0][1]
-        )
-
-        resp2 = resp.follow(
-            extra_environ={'REMOTE_USER': self.container1_admin["name"].encode('ascii')},
-            status=200
-        )
-        orgs = toolkit.get_action("organization_list_for_user")(
-            {"user": self.requesting_user["name"]},
-            {"id": self.requesting_user["name"], "permission": "read"}
-        )
-        assert_equals(self.container1['id'], orgs[0]['id'])
-        assert_equals('approved', self.container1_request.status)
-        assert_in('Access Request Approved', resp2.body)
-
-    def test_access_requests_reject_container_admin(self):
-        mock_mailer = mock.Mock()
-        with mock.patch('ckanext.unhcr.mailer.mail_user_by_id', mock_mailer):
-            resp = self.make_action_request(
-                action='reject',
-                request_id=self.container1_request.id,
-                user=self.container1_admin["name"],
-                status=302,
-                data={'message': 'nope'},
-            )
-        mock_mailer.assert_called_once()
-        # your request has been rejected email
-        assert_equals(
-            '[UNHCR RIDL] - Request for access to: "{}"'.format(self.container1['name']),
-            mock_mailer.call_args[0][1]
-        )
-
-        resp2 = resp.follow(
-            extra_environ={'REMOTE_USER': self.container1_admin["name"].encode('ascii')},
-            status=200
-        )
-        orgs = toolkit.get_action("organization_list_for_user")(
-            {"user": self.requesting_user["name"]},
-            {"id": self.requesting_user["name"], "permission": "read"}
-        )
-        assert_equals(0, len(orgs))
-        assert_equals('rejected', self.container1_request.status)
-        assert_in('Access Request Rejected', resp2.body)
-
-    def test_access_requests_list_invalid_user(self):
-        for user in [None, self.standard_user["name"]]:
-            self.make_list_request(user=user, status=403)
-
-    def test_access_requests_list_sysadmin(self):
-        resp = self.make_list_request(user=self.sysadmin['name'], status=200)
-        # sysadmin can see all the requests
-        assert_in(
-            '/access-requests/approve/{}'.format(self.container1_request.id),
-            resp.body
-        )
-        assert_in(
-            '/access-requests/approve/{}'.format(self.container2_request.id),
-            resp.body
-        )
-        assert_in(
-            '/access-requests/approve/{}'.format(self.dataset_request.id),
-            resp.body
-        )
-        assert_in(
-            '/access-requests/approve/{}'.format(self.user_request_container1.id),
-            resp.body
-        )
-        assert_in(
-            '/access-requests/approve/{}'.format(self.user_request_container2.id),
-            resp.body
-        )
-
-    def test_access_requests_list_container_admin(self):
-        resp = self.make_list_request(user=self.container1_admin['name'], status=200)
-        assert_in(
-            '/access-requests/approve/{}'.format(self.container1_request.id),
-            resp.body
-        )
-        assert_in(
-            '/access-requests/approve/{}'.format(self.dataset_request.id),
-            resp.body
-        )
-        assert_in(
-            '/access-requests/approve/{}'.format(self.user_request_container1.id),
-            resp.body
-        )
-        # container1_admin can't see the requests for container2
-        assert_not_in(
-            '/access-requests/approve/{}'.format(self.container2_request.id),
-            resp.body
-        )
-        assert_not_in(
-            '/access-requests/approve/{}'.format(self.user_request_container2.id),
-            resp.body
-        )
